Move data buffers to ControllerGUIFrame
This is done purely for demonstration purposes, as we are no longer need a Controller object to save buffers to CSV or to manually get a measurement

Now Controller.get_measurements() returns a triple of numbers acquired via serial from the device
